Version Bump to 2.0.0
Do NOT upgrade to this version if you are on 1.x or lower without
carefully ensuring that your application is compatible with this
major version bump.

Version 2.0.0 is _not_ backwards compatible with v1. You must be
very diligent and bump the major version with care.

If it is not an option for you to make sure your application is
compatible with the new version you should remain on 1.x until you
are able to confirm that the new version works for you.

Unicorn Global takes no responsibility for any unintended side
effects that you may encounter if you ignore this disclaimer.